moved RelatedResults logic to new component RelatedResultsField
